Use __wrapped__ to match Python 3.2+
functools.update_wrapper and functools.wraps in Python 3.2+ add a
__wrapped__ attribute to the wrapper function to allow later unwrapping.
Since that's what we're doing in this module, we want to do it in a way
that is consistent with the stdlib.

Since it's not a part of our public interface, changing the name of this
attribute doesn't require any test changes or a major version upgrade.