Fix create_class_wrapper and class_wraps
It appears that none of the attributes in CLASS_WRAPPER_DELETES is
necessary, and two of them (__dict__ and __weakref__) will error out if
they are present, at least in some cases.

To remedy the situation, we add __doc__ to the properties to be copied
over from the Wrapped class, and remove CLASS_WRAPPER_DELETES. Also
remove the corresponding 'deleted' keyword argument to
create_class_wrapper and class_wraps.

To avoid having this come back up again, add a regression test that
makes sure that the wrapper at least doesn't error and copies a few
properties that should work.